refactor: remove distinct types file
